requirements.txt
streamlit
pillow
numpy
scikit-learn
pandas
